requirements.txt
numpy>=1.21.0
pandas>=1.3.0
tensorflow>=2.8.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
pillow>=8.3.0
pathlib 